Add a new bind-mount for vTPM logs
swtpm has its own log directory, and it's hardcoded in libvirt[1].
For the records, this location is currently set to:
/var/log/swtpm/libvirt/qemu

In order to keep some kind of consistency with the current log structure
in /var/log/containers/libvirt, we will keep the "qemu" subdirectory -
it already exists for other qemu-related services, therefore it makes
sense to keep that subdirectory.

This is possible since the swtpm log filename is composed of the
instance ID suffixed by "-swtpm.log", leading to a clear view.

[1] https://gitlab.com/libvirt/libvirt/-/commit/f9cd29a2e44d61da0fc94f245d27206ea66e1161

Related: rhbz#2007314
Related-Bug: #1902468
Change-Id: Ibc80621a622a4eb4ef31520439ded8a08ce50c96