Remove stack_update check in rabbit manifest
Since we have ephemeral heat the stack_update hiera key is always
set to CREATE anyway so it will never work.

We can drop it because we have the puppet collectors running that
after rabbitmq_ready so rabbitmq is up in the HA case.

For the non-HA case (rabbitmq on the undercloud) we run rabbitmq
via docker_config at step_1 and then enforce the rabbitmq_user class
during container_puppet_tasks (rabbitmq_init_tasks) during step 2,
so we're guaranteed that rabbitmq is up.

Change-Id: If3d16b8510175ed9fae2813fbdc7248402639e78