Bind mount the IPA crt when internal_tls is enabled
In order for later reviews to make use of the FreeIPA internal
CA we need to first bind mount it within the container.

We need to add a default in the hiera definition (/etc/ipa/ca.crt)
in order to break a cyclic dependency on the subsequent patches.
(THT child change will set the rabbitmq::ssl_cacert key)

Related-Bug: #1946374
Change-Id: Ib0236f9c086d520d0a27e3aa8b41927bc7b50c26