Make sure resource_op_defaults are set before bundles
See LP#1948032. We set a higher default resource op time out because
podman often times out within the default 20s. We need to guarantee
that this timeout is set *before* we create any bundles otherwise
it might be too late and we'd fail a deployment like described in the
LP.

Change-Id: Ic6ac8d21d12389e70cd382018290324d3ae948ce
Closes-Bug: #1948032