Remove ineffective unit test code
This is a follow-up of 65e7e338f4872f43f35d2ba57666499ce45dbcfe and
removes the unit test code added by that change, because the code
actually invokes no tests. Also, the parameter it intends to test is
not a puppet parameter and we can't test this by overriding params.

Change-Id: I2b6c32073151d793704601d4da64fd7c47fee711