Enable policy rule management in nova-compute
There are some features in nova-compute which validate policy rules.
For example when connecting an instance to an external network,
nova-compute checks ``network:attach_external_network`` to determine
whether the operation is permitted.

This change makes sure that the nova policy file in compute nodes are
also managed by puppet-tripleo.

Partial-Bug: #1955786
Change-Id: I490cc558238719d4c9585e2a57497d1b1787a9ed