Format [keystone_authtoken] memcache_servers when IPv6 is used
When Memcached uses IPv6 network, the memcache_servers parameter should
be formatted as is described in the following example.
 inet6:[<host>]:<port>

This change ensures the proper format is applied even when hostnames
are used to define servers.

Also, this change fixes the timing to apply any2array. The function
should be applied before we check the first memcache server by [0],
otherwise the logic to detect IPv6 address does not work as intended.

Partial-Bug: #1964824
Depends-on: https://review.opendev.org/833708
Change-Id: I18537ab819996cfb0f2705d8c49666d4b9bfff22