Format [oslo_cache] memcache_server when IPv6 is used
When Memcached uses IPv6 network, python-memcached requires that each
server name is formatted as is described in the following example.
 inet6:[<host>]:<port>

This change ensures the format is properly applied according to
the IP protocol version and cache backend used.

Note that the parameter in keystone was not properly formatted even
when IPs are used to set the parameter. This change fixes that and
ensure the parameter is properly configured.

Also, this change fixes the timing to apply any2array. The function
should be applied before we check the first memcache server by [0],
otherwise the logic to detect IPv6 address does not work as intended.

Closes-Bug: #1964824
Change-Id: I22f8fc7f59b4eeac10c3a274c36daeaa1861fd69