Manila: Remove unnecessary 'undef'
The third argument of the hiera function determines the hierarchy level
to insert at the top of the hierarchy[1].
 [1] https://puppet.com/docs/puppet/7/function.html#hiera

This is currently used for dellemc_unity::emc_share_backend only and is
not used in the other places. In addition passing undef has no effect
so it is just redundant.

This change removes the unnecessary undef parameter, to make usage of
the hiera function consistent across all manifests.

Change-Id: Ia59a79fbd293775afdacd9f8a517ff256720a7c0