Remove package update by puppet
The tripleo::packages::enable_upgrade parameter allows updating host
packages by puppet. However this feature is no longer used since we
introduced update tasks mechanism and usage of this parameter is quite
likely to cause a problem.

Because usage of this parameter is highly discouraged now, and we do
not expect any user still relying on this old mechanism, this change
removes the parameter.

Change-Id: I7558b4a5398d0c3a87aecc6017458a5a5932c644