[Train ONLY] Update local certmonger ca cert when ca cert has been renewed Resues fix from "Always update the local certmonger ca cert" https://review.opendev.org/c/openstack/tripleo-heat-templates/+/785020
This fix would not work in wallby up as this was moved 
to tripleo-ansible.
From review the code in tripleo-ansible, I am not sure how we would vaildate that the
certificate was issued by the local certmonger CA. 
This is important as issuing certificates from local certmonger CA is only done on the 
undercloud.
But the role is used for the both undercloud and overcloud.

Closes-Bug: 1988244
Change-Id: I3dd3b9c0aa51ffa5b69a69d4bf113ae41eeffd81