Template option addition for --pids-limit on Galera cluster resource
mysql_bundle.pp has been modified to be able to configure --pids-limit option
using a template. By default the parameter remains 'undef' when not specified.
When "tripleo::profile::pacemaker::database::mysql_bundle::pids_limit:" is used
as an ExtraConfig it will automatically set the value in the Galera cluster
resource.

Closes-Bug: #1982751

Change-Id: Iffe9c9a1d7ca736f273d2da43928d7da4a99d1d6